Proof of Concept Dog Update
Added dog to scene
Made states for the dog depending on it's mood and distance to the player.
Using the bone on the dog will make it happy (showed by turning pink)
Punching the dog will make it angry (showed by turning red).
If it's not angry, the dog will stop moving towards you when it's close.

Current bugs:
Dog is able to be pushed by the player moving into it.
Player sometimes ends up on top of dog from collisions.  (possibly due to capsule colliders)