Gameplay Update 1
Added Skeleton (with placeholder model) that can be killed with the Axe to get a Bone

Readded Placeholder Dog and changed its behavior so that it blocks the Manor entrance unless Angered, Killed, or Befriended

Add the punch and grab animations to the in game model.  Controls are disabled while those animations play.